fibonacci

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -11,8 +11,8 @@
 int main(void)
 {
     int i;
-    long long j;
-    long long k;
+    long j;
+    long k;
 
     i = 0; 
     j = 1;
@@ -21,14 +21,14 @@ int main(void)
     while (i < 50)
     {
         if (i == 0)
-            printf("%lli", j);
+            printf("%li", j);
         else if (i == 1)
-            printf("%lli", k);
+            printf("%li", k);
         else 
         {
             k += j;
             j = k - j;
-            printf("%lli", k);
+            printf("%li", k);
         }
 
         if (i != 49)
